UserLayout , UserBlogList , UserCommentList
UserLayout güncellendi

Usera göre blogların listelenmesi yapıldı

usera göre yorumlarının listelenmesi yapıldı

+++ BugsByteLibrary/Areas/Admin/Controllers/RoleController.cs
@@ -7,7 +7,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace BugsByteLibrary.Areas.Admin.Controllers
 {
-    [Authorize(Roles ="Admin")]
+    
     [Area("Admin")]
     public class RoleController : Controller
     {
@@ -22,6 +22,11 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> Index()
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             var value = await _roleManager.Roles.ToListAsync();
             return View(value);
         }
@@ -29,6 +34,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpGet]
         public IActionResult AddRole()
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
             return View();
         }
 
@@ -36,6 +45,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> AddRole(CreateRoleViewModel createRoleViewModel)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
             AppRole role = new()
             {
                 Name = createRoleViewModel.RoleName
@@ -56,6 +69,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         }
         public async Task<IActionResult>DeleteRole(int id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
             var role =  await _roleManager.Roles.FirstOrDefaultAsync(x => x.Id == id);
             await _roleManager.DeleteAsync(role);
 
@@ -66,6 +83,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpGet]
         public async Task<IActionResult> EditRole(int id )
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
             var role = await _roleManager.Roles.FirstOrDefaultAsync(x => x.Id == id);
             EditRoleViewModel editRoleViewModel = new()
             {
@@ -80,6 +101,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> EditRole(EditRoleViewModel editRoleViewModel)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
             var value =  _roleManager.Roles.FirstOrDefault(x=>x.Id == editRoleViewModel.RoleId);
             value.Name = editRoleViewModel.RoleName;
             value.Id = value.Id;
@@ -93,12 +118,20 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> UserList()
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
             var value = await _userManager.Users.ToListAsync();
             return View(value);
         }
 
         public async Task<IActionResult> GiveRole(int id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
             var userGetById = await _userManager.Users.FirstOrDefaultAsync(x=> x.Id == id);
             TempData["UserId"] = userGetById.Id;
             var roles = await _roleManager.Roles.ToListAsync();
@@ -118,6 +151,11 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> GiveRole(List<GiveRoleViewModel> models)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             var userId = (int)TempData["UserId"];
             var user = await _userManager.Users.FirstOrDefaultAsync(x=> x.Id ==userId );
